dev: refactor device tests and add http api (#233)
This PR refactors the online device tests, end to end API tests and adds offline device tests. Also, this PR refactors the API implementation to decouple the endpoint type with handlers.

+++ iwfpapi/functions/src/handler/remove_credit_card.ts
@@ -1,13 +1,24 @@
 import { noAuthMsg } from "../config/consts";
 import getUserUid from "../util/uid_getter";
 import isValidAuth from "../util/validate_auth";
+import { CardRemovalRequest } from "../config/typedefs";
+import { CardForDeletionNotExist } from "../config/errors";
 
-async function removeCreditCardHandler(data, context, provider) {
+async function removeCreditCardHandler(
+  data: CardRemovalRequest,
+  context,
+  provider
+) {
   if (isValidAuth(context, process.env.FUNCTIONS_EMULATOR)) {
     const userUid: string = getUserUid(context, process.env.FUNCTIONS_EMULATOR);
     const userRef = provider.getUserRef(userUid);
-    const cardRef = userRef.collection("cards").doc(data.cardUid);
-    await cardRef.delete();
+    const cardRef = userRef.collection("cards").doc(data.id);
+    const cardSnap = await cardRef.get();
+    if (cardSnap.exists) {
+      await cardRef.delete();
+    } else {
+      throw CardForDeletionNotExist;
+    }
   } else {
     throw noAuthMsg;
   }
